Implement basic Moneypress eBay plugin and shortcode
Here we implement a basic version of the plugin along with a
configuration file that loads the appropriate CSL sub-modules.  This
commit _does not_ work as-is because it relies on getting the config
value 'csl-mp-ebay-app-id', and there is currently no way to set that
value except by changing the code by hand.  An up-coming commit will
introduce admin settings for this plugin which will allow users to set
that value.

The display of the products is also very limited right now.  It is
nothing but a list of the titles of results, where the titles are
links to pages where the items can be purchased.

Redmine-Ticket: 2066
Redmine-Ticket: 2070