Create admin settings page
This uses the generic settings sub-module to provide a place to enter
the license key and display information about Cyber Sprocket as a
company.  But it also includes two plugin-specific settings for
entering a developer application key, which is needed to fetch
products, and a field to set the number of products to display.

Redmine-Ticket: 2068