Remove some of the white-space inside of product listings
Currently there is a large empty space to the right of the images
where normally a product description would be displayed.  Right now we
do not get that information out of the eBay driver, so for the time
being we move the price into that space and cut down on the vertical
size of the product lists.

We also fix the spelling of ‘yeti’.

Redmine-Ticket: 2071