Teach plugin to work with the new WPCSL-Generic submodule
The new submodule provides some convenient functions that encapsulate
some of the existing behavior in the plugin.  Making the switch only
requires us to replace that existing behavior with calls to the
appropriate methods of the wpCSL_plugin class.

Since we now instantiate that instead of a settings class, the
variable $MP_ebay_settings was renamed everywhere to be
$MP_ebay_plugin.

Redmine-Ticket: 2093