Update the eBay driver to have better error handling
Previously the eBay driver would die if it were missing any
requirements, but now it throws an exception in that case.  Instead of
causing Wordpress to die, we now catch that exception and post a
notification on the admin page informing the the user that he is
missing a required library.

Redmine-Ticket: 2112