Allow the shortcode to be used without keywords
If a user does not provide any keywords then we show all of the
products associated with the seller ID they provide from the admin
panel.  It is still possible to use keywords and restrict the results
to a specific vendor as well.

This patch includes an update to the Panhandler eBay driver so that we
have access to the new get_products_from_vendor() method.

During the course of this work, a bug came up in a situation where
the number of products to display could be zero.  The code which
introduced that bug has been removed from this commit, but comments
have been added in the code explaining that we have to guard against
that situation.

Redmine-Ticket: 2122