Pass affiliate tracking ID to eBay when finding products
The way we give affiliate information to the eBay driver is via the
same mechanism that we use to pass the seller ID when we are fetching
products by keywords.  So this gives us a chance to refactor some
code.  The existing code uses a series of if-else statements to decide
what to pass as the second argument to get_products_from_vendor() or
get_products_by_keywords().  Now we use a new function to create the
value for that argument and then pass that in regardless of any other
logic.  This is a safer operation since Panhandler drivers are
required to ignore options they don’t understand, so we’re not going
to trip up any method here by passing in the wrong set of options.

Redmine-Ticket: 2113