Allow users to enter their partner network ID
This is used to associate affiliate commissions between eBay and the
user.  This field is required if the user provides a tracking ID.
Currently we do not enforce that requirement on the user-interface,
i.e. we do not show them a warning if they enter one without the
other, but we do enforce it in the logic at a code level.

Redmine-Ticket: 2130