Set WPCSL-Generic to the proper commit for the v1.2 release
The eBay plugin is currently not written to take advantage of the
latest changes to the plugin framework.  It relies on the initial
implementation of list settings, which has not yet made it into the
mainline for the framework.  Therefore we need to set the sub-project
a few commits back from the latest work in order for the eBay plugin
to function properly.